HPCC-34696 Move to WsSMC
Add verification to require DateDimeRange on request. If not present response
will be empty.
