+++ esp/services/ws_machine/ws_machineService.cpp
@@ -25,7 +25,6 @@
 #include "rmtssh.hpp"
 #include "platform.h"
 #include "TpWrapper.hpp"
-#include "dastats.hpp"
 
 #include <math.h>
 
@@ -119,41 +118,6 @@ public:
     {
         m_pService->getMachineUsage(espContext, this);
     }
-};
-
-class CArrayGlobalMetricRecorder : implements IGlobalMetricRecorder
-{
-public:
-    CArrayGlobalMetricRecorder(IArrayOf<IEspGlobalMetric> & _target) : target(_target) {}
-    
-    void processGlobalStatistics(const char * category, const MetricsDimensionList & dimensions, const char * startTime, const char * endTime, const GlobalStatisticsList & stats) override
-    {
-        Owned<IEspGlobalMetric> targetMetric = createGlobalMetric();
-        targetMetric->setCategory(category);
-        targetMetric->updateDateTimeRange().setStart(startTime);
-        targetMetric->updateDateTimeRange().setEnd(endTime);
-
-        for (const auto & dimension : dimensions)
-        {
-            Owned<IEspDimension> targetDimension = createDimension();
-            targetDimension->setName(dimension.first);
-            targetDimension->setValue(dimension.second);
-            targetMetric->getDimensions().append(*targetDimension.getClear());
-        }
-        
-        for (const auto & stat : stats)
-        {
-            Owned<IEspStat> targetStat = createStat();
-            targetStat->setName(queryStatisticName(stat.first));
-            targetStat->setValue(std::to_string(stat.second).c_str());
-            targetMetric->getStats().append(*targetStat.getClear());
-        }
-
-        target.append(*targetMetric.getClear());
-    }
-
-private:
-    IArrayOf<IEspGlobalMetric> & target;
 };
 
 const char* findComponentTypeFromProcessType(const char* ProcessType)
@@ -342,43 +306,7 @@ bool Cws_machineEx::onGetTargetClusterInfo(IEspContext &context, IEspGetTargetCl
     return true;
 }
 
-void handleGetGlobalMetrics(IEspGetGlobalMetricsRequest &req, IEspGetGlobalMetricsResponse &resp)
-{
-    const char* category = req.getCategory();
-    MetricsDimensionList dimensions;
-    ForEachItemIn(i, req.getDimensions())
-    {
-        IConstDimension& dim = req.getDimensions().item(i);
-        if (!isEmptyString(dim.getName()))
-            dimensions.push_back({dim.getName(), dim.getValue()});
-    }
 
-    if (isEmptyString(category) || strieq(category, "All"))
-        category = nullptr;
-
-    CDateTime start, end;
-    start.setString(req.getDateTimeRange().getStart(), NULL, false);
-    end.setString(req.getDateTimeRange().getEnd(), NULL, false);
-    IArrayOf<IEspGlobalMetric> metricsResponse;
-    CArrayGlobalMetricRecorder metricRecorder(metricsResponse);
-    gatherGlobalMetrics(category, dimensions, start, end, metricRecorder);
-
-    resp.setGlobalMetrics(metricsResponse);
-}
-
-bool Cws_machineEx::onGetGlobalMetrics(IEspContext &context, IEspGetGlobalMetricsRequest &req, IEspGetGlobalMetricsResponse &resp)
-{
-    try
-    {
-        context.ensureFeatureAccess(FEATURE_URL, SecAccess_Read, ECLWATCH_MACHINE_INFO_ACCESS_DENIED, "Failed to get global metrics. Permission denied.");
-        handleGetGlobalMetrics(req, resp);
-    }
-    catch (IException *e)
-    {
-        FORWARDEXCEPTION(context, e, ECLWATCH_INTERNAL_ERROR);
-    }
-    return true;
-}
 
 void Cws_machineEx::addChannels(CGetMachineInfoData& machineInfoData, IPropertyTree* envRoot,
     const char* componentType, const char* componentName)
@@ -3565,213 +3493,3 @@ void CUsageCacheReader::addOtherComponentUsageReq(IConstEnvironment* constEnv, c
         checkAndAddMachineUsageReq(constEnv, computer, logFolder, dataFolder, repFolder, usageReq);
     }
 }
-
-#ifdef _USE_CPPUNIT
-//---------------------------------------------------------------------------
-// Unit test for handleGetGlobalMetrics() local helper.
-// Name prefixed with 'Dali' to indicate it needs a locally-running dali
-// and therefore automatically excluded from automated tests.
-//---------------------------------------------------------------------------
-
-#include "mpbase.hpp"
-#include "mpcomm.hpp"
-#include "daclient.hpp"
-#include "dasds.hpp"
-#include "dautils.hpp"
-#include <cppunit/extensions/HelperMacros.h>
-#include <algorithm>
-
-// copy daliClientInit/daliClientEnd approach from testing/unittests/dalitests.cpp
-static unsigned wsMachineDaliInitCounter = 0;
-static Owned<IUserDescriptor> wsMachineUser;
-
-static void daliClientInit()
-{
-    if (wsMachineDaliInitCounter != 0)
-        return;
-    InitModuleObjects();
-    wsMachineUser.setown(createUserDescriptor());
-    wsMachineUser->set("user", "passwd");
-    SocketEndpoint ep;
-    ep.set(".", 7070); // local dali (standard test harness port)
-    SocketEndpointArray epa;
-    epa.append(ep);
-    Owned<IGroup> group = createIGroup(epa);
-    initClientProcess(group, DCR_Testing);
-    wsMachineDaliInitCounter++;
-}
-
-static void daliClientEnd()
-{
-    if (!wsMachineDaliInitCounter)
-        return;
-    else if (wsMachineDaliInitCounter == 1)
-    {
-        releaseAtoms();
-        closedownClientProcess();
-        setNodeCaching(false);
-        wsMachineUser.clear();
-        wsMachineDaliInitCounter = 0;
-    }
-    else
-        wsMachineDaliInitCounter--;
-}
-
-
-class DaliWsMachineGlobalMetricsTest : public CppUnit::TestFixture
-{
-    CPPUNIT_TEST_SUITE(DaliWsMachineGlobalMetricsTest);
-        CPPUNIT_TEST(setUp);
-        CPPUNIT_TEST(testAllCategoriesExplicit);
-        CPPUNIT_TEST(testAllCategoriesImplicit);
-        CPPUNIT_TEST(testSingleCategoryAndDimensions);
-        CPPUNIT_TEST(testDimensionAcrossCategories);
-        CPPUNIT_TEST(tearDown);
-    CPPUNIT_TEST_SUITE_END();
-
-public:
-    void setUp() override
-    {
-        daliClientInit();
-        resetGlobalMetrics("categoryOne", MetricsDimensionList());
-        resetGlobalMetrics("categoryTwo", MetricsDimensionList());
-        setGlobalMetricNowTime("1999-07-01T12:00:00");
-        prepareSampleMetrics();
-    }
-    void tearDown() override
-    {
-        daliClientEnd();
-    }
-
-    void prepareSampleMetrics()
-    {
-        recordGlobalMetrics("categoryOne", MetricsDimensionList{}, { StTimeLocalExecute }, { 111 });
-        recordGlobalMetrics("categoryOne", MetricsDimensionList{{"user","alice"}}, { StTimeLocalExecute, StCostExecute }, { 222, 5 });
-        recordGlobalMetrics("categoryOne", MetricsDimensionList{{"user","bob"},{"cluster","thor1"}}, { StTimeLocalExecute }, { 333 });
-        recordGlobalMetrics("categoryTwo", MetricsDimensionList{}, { StTimeLocalExecute }, { 444 });
-        recordGlobalMetrics("categoryTwo", MetricsDimensionList{{"user","alice"}}, { StTimeLocalExecute, StCostExecute }, { 555, 7 });
-        recordGlobalMetrics("categoryTwo", MetricsDimensionList{{"user","bob"},{"cluster","thor1"}}, { StTimeLocalExecute }, { 666 });
-    }
-
-    Owned<IEspGetGlobalMetricsRequest> createRequestFromXML(const char * xml)
-    {
-        Owned<IPropertyTree> t = createPTreeFromXMLString(xml);
-        Owned<IEspGetGlobalMetricsRequest> req = createGetGlobalMetricsRequest();
-        if (const char *cat = t->queryProp("Category"))
-            req->setCategory(cat);
-        if (IPropertyTree * rng = t->queryPropTree("DateTimeRange"))
-        {
-            if (const char *s = rng->queryProp("Start"))
-                req->updateDateTimeRange().setStart(s);
-            if (const char *e = rng->queryProp("End"))
-                req->updateDateTimeRange().setEnd(e);
-        }
-        Owned<IPropertyTreeIterator> dims = t->getElements("Dimensions/Dimension");
-        ForEach(*dims)
-        {
-            IPropertyTree &d = dims->query();
-            const char *name = d.queryProp("Name");
-            if (isEmptyString(name))
-                continue;
-            Owned<IEspDimension> dim = createDimension();
-            dim->setName(name);
-            dim->setValue(d.queryProp("Value"));
-            req->getDimensions().append(*dim.getClear());
-        }
-        return req.getClear();
-    }
-
-    void serializeResponse(IEspGetGlobalMetricsResponse &resp, StringBuffer &out)
-    {
-        // Use generated serializer from ws_machine.esp for canonical XML
-        CGetGlobalMetricsResponse *impl = dynamic_cast<CGetGlobalMetricsResponse *>(&resp);
-        CPPUNIT_ASSERT_MESSAGE("Response object type unexpected", impl != nullptr);
-        CGetGlobalMetricsResponse::serializer(nullptr, *impl, out, true); // keep root tag
-    }
-
-    static constexpr const char * expectedAllCategories = "<GetGlobalMetricsResponse><GlobalMetrics>"
-        "<GlobalMetric><Category>categoryTwo</Category><DateTimeRange><Start>1999070112</Start><End>1999070112</End></DateTimeRange><Stats><Stat><Name>TimeLocalExecute</Name><Value>444</Value></Stat></Stats></GlobalMetric>"
-        "<GlobalMetric><Category>categoryTwo</Category><Dimensions><Dimension><Name>user</Name><Value>alice</Value></Dimension></Dimensions><DateTimeRange><Start>1999070112</Start><End>1999070112</End></DateTimeRange><Stats><Stat><Name>TimeLocalExecute</Name><Value>555</Value></Stat><Stat><Name>CostExecute</Name><Value>7</Value></Stat></Stats></GlobalMetric>"
-        "<GlobalMetric><Category>categoryTwo</Category><Dimensions><Dimension><Name>user</Name><Value>bob</Value></Dimension><Dimension><Name>cluster</Name><Value>thor1</Value></Dimension></Dimensions><DateTimeRange><Start>1999070112</Start><End>1999070112</End></DateTimeRange><Stats><Stat><Name>TimeLocalExecute</Name><Value>666</Value></Stat></Stats></GlobalMetric>"
-        "<GlobalMetric><Category>categoryOne</Category><DateTimeRange><Start>1999070112</Start><End>1999070112</End></DateTimeRange><Stats><Stat><Name>TimeLocalExecute</Name><Value>111</Value></Stat></Stats></GlobalMetric>"
-        "<GlobalMetric><Category>categoryOne</Category><Dimensions><Dimension><Name>user</Name><Value>alice</Value></Dimension></Dimensions><DateTimeRange><Start>1999070112</Start><End>1999070112</End></DateTimeRange><Stats><Stat><Name>TimeLocalExecute</Name><Value>222</Value></Stat><Stat><Name>CostExecute</Name><Value>5</Value></Stat></Stats></GlobalMetric>"
-        "<GlobalMetric><Category>categoryOne</Category><Dimensions><Dimension><Name>user</Name><Value>bob</Value></Dimension><Dimension><Name>cluster</Name><Value>thor1</Value></Dimension></Dimensions><DateTimeRange><Start>1999070112</Start><End>1999070112</End></DateTimeRange><Stats><Stat><Name>TimeLocalExecute</Name><Value>333</Value></Stat></Stats></GlobalMetric>"
-        "</GlobalMetrics></GetGlobalMetricsResponse>";
-
-    void testAllCategoriesExplicit()
-    {
-        static const char * requestXML =
-            "<GetGlobalMetricsRequest>"
-              "<Category>All</Category>"
-              "<DateTimeRange><Start>1999-01-01T00:00:00</Start><End>2099-01-01T00:00:00</End></DateTimeRange>"
-            "</GetGlobalMetricsRequest>";
-        Owned<IEspGetGlobalMetricsRequest> req = createRequestFromXML(requestXML);
-        Owned<IEspGetGlobalMetricsResponse> resp = createGetGlobalMetricsResponse();
-        handleGetGlobalMetrics(*req, *resp);
-        StringBuffer out; 
-        serializeResponse(*resp, out);
-    CPPUNIT_ASSERT_EQUAL(std::string(expectedAllCategories), std::string(out.str()));
-    }
-
-    void testAllCategoriesImplicit()
-    {
-        static const char * requestXML =
-            "<GetGlobalMetricsRequest>"
-              "<DateTimeRange><Start>1999-01-01T00:00:00</Start><End>2099-01-01T00:00:00</End></DateTimeRange>"
-            "</GetGlobalMetricsRequest>";
-        Owned<IEspGetGlobalMetricsRequest> req = createRequestFromXML(requestXML);
-        Owned<IEspGetGlobalMetricsResponse> resp = createGetGlobalMetricsResponse();
-        handleGetGlobalMetrics(*req, *resp);
-        StringBuffer out; 
-        serializeResponse(*resp, out);
-    CPPUNIT_ASSERT_EQUAL(std::string(expectedAllCategories), std::string(out.str()));
-    }
-
-    void testSingleCategoryAndDimensions()
-    {
-        static const char * requestXML =
-            "<GetGlobalMetricsRequest>"
-              "<Category>categoryOne</Category>"
-              "<Dimensions>"
-                "<Dimension><Name>user</Name><Value>bob</Value></Dimension>"
-              "</Dimensions>"
-              "<DateTimeRange><Start>1999-01-01T00:00:00</Start><End>2099-01-01T00:00:00</End></DateTimeRange>"
-            "</GetGlobalMetricsRequest>";
-        Owned<IEspGetGlobalMetricsRequest> req = createRequestFromXML(requestXML);
-        Owned<IEspGetGlobalMetricsResponse> resp = createGetGlobalMetricsResponse();
-        handleGetGlobalMetrics(*req, *resp);
-        StringBuffer out; 
-        serializeResponse(*resp, out);
-        static const char * expected =
-            "<GetGlobalMetricsResponse><GlobalMetrics>"
-                "<GlobalMetric><Category>categoryOne</Category><Dimensions><Dimension><Name>user</Name><Value>bob</Value></Dimension><Dimension><Name>cluster</Name><Value>thor1</Value></Dimension></Dimensions><DateTimeRange><Start>1999070112</Start><End>1999070112</End></DateTimeRange><Stats><Stat><Name>TimeLocalExecute</Name><Value>333</Value></Stat></Stats></GlobalMetric>"
-            "</GlobalMetrics></GetGlobalMetricsResponse>";
-        CPPUNIT_ASSERT_EQUAL(std::string(expected), std::string(out.str()));
-    }
-
-    void testDimensionAcrossCategories()
-    {
-        static const char * requestXML =
-            "<GetGlobalMetricsRequest>"
-              "<Dimensions>"
-                "<Dimension><Name>user</Name><Value>alice</Value></Dimension>"
-              "</Dimensions>"
-              "<DateTimeRange><Start>1999-01-01T00:00:00</Start><End>2099-01-01T00:00:00</End></DateTimeRange>"
-            "</GetGlobalMetricsRequest>";
-        Owned<IEspGetGlobalMetricsRequest> req = createRequestFromXML(requestXML);
-        Owned<IEspGetGlobalMetricsResponse> resp = createGetGlobalMetricsResponse();
-        handleGetGlobalMetrics(*req, *resp);
-        StringBuffer out; 
-        serializeResponse(*resp, out);
-        static const char * expected =
-            "<GetGlobalMetricsResponse><GlobalMetrics>"
-              "<GlobalMetric><Category>categoryTwo</Category><Dimensions><Dimension><Name>user</Name><Value>alice</Value></Dimension></Dimensions><DateTimeRange><Start>1999070112</Start><End>1999070112</End></DateTimeRange><Stats><Stat><Name>TimeLocalExecute</Name><Value>555</Value></Stat><Stat><Name>CostExecute</Name><Value>7</Value></Stat></Stats></GlobalMetric>"
-              "<GlobalMetric><Category>categoryOne</Category><Dimensions><Dimension><Name>user</Name><Value>alice</Value></Dimension></Dimensions><DateTimeRange><Start>1999070112</Start><End>1999070112</End></DateTimeRange><Stats><Stat><Name>TimeLocalExecute</Name><Value>222</Value></Stat><Stat><Name>CostExecute</Name><Value>5</Value></Stat></Stats></GlobalMetric>"
-            "</GlobalMetrics></GetGlobalMetricsResponse>";
-        CPPUNIT_ASSERT_EQUAL(std::string(expected), std::string(out.str()));
-    }
-};
-
-CPPUNIT_TEST_SUITE_REGISTRATION(DaliWsMachineGlobalMetricsTest);
-CPPUNIT_TEST_SUITE_NAMED_REGISTRATION(DaliWsMachineGlobalMetricsTest, "DaliWsMachineGlobalMetricsTest");
-#endif // _USE_CPPUNIT
